Fix post unidad organizativa

+++ src/SGVO.Infrastructure/Commands/EliminarTipoUnidadOrganizativaCommandHandler.cs
@@ -24,7 +24,7 @@ public class EliminarTipoUnidadOrganizativaCommandHandler : ICommandHandler<Elim
         CancellationToken cancellationToken)
     {
         var entity = await _dbContext.TiposUnidadOrganizativa
-            .FirstOrDefaultAsync(e => e.Id == (ulong)command.Id, cancellationToken);
+            .FirstOrDefaultAsync(e => e.Id == command.Id, cancellationToken);
 
         if (entity is null)
             return Result<Unit>.Failure(
