+++ tests/SGVO.UnitTests/Commands/TiposUnidadOrganizativa/EliminarTipoUnidadOrganizativaCommandHandlerTests.cs
@@ -36,7 +36,7 @@ public class EliminarTipoUnidadOrganizativaCommandHandlerTests
         var entity = dbContext.TiposUnidadOrganizativa.First(e => e.Id == 1);
         Assert.False(entity.Activo);
         Assert.NotNull(entity.EliminadoEn);
-        Assert.Equal((ulong)42, entity.EliminadoPor);
+        Assert.Equal(42L, entity.EliminadoPor);
     }
 
     [Fact]
